+++ Providers/EasyNetQ/Revo.EasyNetQ/Configuration/EasyNetQConnectionConfiguration.cs
@@ -1,0 +1,12 @@
+﻿namespace Revo.EasyNetQ.Configuration
+{
+    public class EasyNetQConnectionConfiguration
+    {
+        public EasyNetQConnectionConfiguration(string connectionString)
+        {
+            ConnectionString = connectionString;
+        }
+
+        public string ConnectionString { get; }
+    }
+}