+++ Providers/EasyNetQ/Revo.EasyNetQ/IEasyNetQSubscriptionHandler.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace Revo.EasyNetQ
+{
+    public interface IEasyNetQSubscriptionHandler
+    {
+        Task HandleMessage(object message);
+    }
+}